teleruptor: drop fail state
The number of automatic attempts to get the teleruptor in the right
state is limited, to avoid endless activation of failing devices.

But the teleruptor driver will never go into FAIL state.
The driver returns to IDLE rather than FAIL.
As such, the device's condition is not really correct, but another
attempt to correct is tried upon setting a new value. This latter
occurs on user input.