motor: add EOL control options
Control the EOL behaviour, and for both ends to behave differently.
The old behaviour is the default (both EOL0 & EOL1).

This patch enables the use of a window-open motor.
The 0 position (window closed) may expose some extra _make sure it's
closed_ time, and the 1 position (window open) should not.
The old behaviour (Jalousie) is still also valid: _make sure it's
completely opened/closed_.