hamotor: fix direction switch
Before this commit, when a switch is pressed after the first motor
has finished, but while a second is still moving, the motors will
both start in reverse.
The expected behaviour is to stop all motors first.
When no motor is moving, then a buttonpress will engage motors
in reverse direction.

Signed-off-by: Kurt Van Dijck <dev.kurt@vandijck-laurijssen.be>